+++ lib/WorseReflection/Tests/Integration/Core/Inference/NodeContextResolverTest.php
@@ -5,6 +5,7 @@ namespace Phpactor\WorseReflection\Tests\Integration\Core\Inference;
 use Phpactor\WorseReflection\Bridge\Phpactor\DocblockParser\DocblockParserFactory;
 use Phpactor\WorseReflection\Core\Cache\StaticCache;
 use Phpactor\WorseReflection\Core\DefaultResolverFactory;
+use Phpactor\WorseReflection\Core\Inference\Frame\ConcreteFrame;
 use Phpactor\WorseReflection\Core\Inference\GenericMapResolver;
 use Phpactor\WorseReflection\Core\Inference\NodeToTypeConverter;
 use Phpactor\WorseReflection\Core\Inference\PropertyAssignments;
@@ -13,7 +14,6 @@ use Phpactor\WorseReflection\Core\Inference\Resolver\MemberAccess\NodeContextFro
 use Phpactor\WorseReflection\Core\TypeFactory;
 use Phpactor\WorseReflection\Tests\Integration\IntegrationTestCase;
 use Phpactor\WorseReflection\Core\Type;
-use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Inference\LocalAssignments;
 use Phpactor\WorseReflection\Core\Inference\Variable;
 use Phpactor\WorseReflection\Core\Inference\Symbol;
@@ -1166,7 +1166,7 @@ class NodeContextResolverTest extends IntegrationTestCase
         PropertyAssignments $properties,
         string $source
     ): NodeContext {
-        $frame = new Frame($locals, $properties);
+        $frame = new ConcreteFrame($locals, $properties);
 
         [$source, $offset] = ExtractOffset::fromSource($source);
         $node = $this->parseSource($source)->getDescendantNodeAtPosition($offset);